+++ src/core/pullRequests.ts
@@ -11,6 +11,7 @@ interface FetchPullRequests {
   endDate: string;
   outputDirectory: string;
   logFilePath: string;
+  capture: boolean;
 }
 
 interface GitHubApiResponse {
@@ -23,6 +24,7 @@ export const fetchPullRequests = async ({
   endDate,
   outputDirectory,
   logFilePath,
+  capture,
 }: FetchPullRequests) => {
   console.log(
     `${successColorAnsi("[*]")} Fetching data from author: ${warningColorAnsi(
@@ -84,5 +86,5 @@ export const fetchPullRequests = async ({
     fs.mkdirSync(outputDirectory, { recursive: true });
   }
 
-  sequentialProcess(data, outputDirectory, logFilePath);
+  sequentialProcess(data, outputDirectory, logFilePath, capture);
 };